Clean up the code a bit. Remove TFCameraControls, TFGraphCanvas, TFGraphScene, TFNode, TFRing. Keeping in mind that reparated PRs with fixes should be merge first. Move TFCustomNode, TFSphereWithBorder to /custom/ namespace.
