+++ src/CameraControls/index.ts
@@ -1,4 +1,3 @@
 export * from './CameraControls';
 export * from './useCameraControls';
 export * from './useCenterGraph';
-export * from './TFCameraControls';
