+++ src/index.ts
@@ -11,5 +11,4 @@ export * from './RadialMenu';
 export * from './CameraControls';
 export * from './useGraph';
 export * from './themes';
-export * from './TFGraphCanvas';
-export * from './TFGraphScene';
+export * from './custom';
