+++ src/symbols/index.ts
@@ -7,5 +7,3 @@ export * from './edges';
 export * from './nodes';
 export * from './Ring';
 export * from './Cluster';
-export * from './TFNode';
-export * from './TFRing';
