+++ src/symbols/nodes/index.ts
@@ -1,7 +1,5 @@
 export * from './Icon';
 export * from './Sphere';
+export * from './SphereWithIcon';
 export * from './Svg';
 export * from './SphereWithSvg';
-export * from './SphereWithIcon';
-export * from './TFCustomNode';
-export * from './TFSphereWithBorder';
